+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Chore
 
 - Improve image input/output API.
+- Groovy syntax accesses image input/output features on container extension. 
 
 ## [0.2.1] - 2025-04-23